added function comments
